# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: COMPAON API
  version: 1.0.0
  contact:
    email: compaon@dalavina.com
    name: Compaon
    url: 'https://dalavina.com/compaon'
  license:
    name: EULA
    url: 'https://en.wikipedia.org/wiki/End-user_license_agreement'
  description: API definition for COMPAON
servers:
  - url: 'http://52.74.156.150:3000/api/v1'
paths:
...
  '/booths/{id}':
    get:
      security:
      - bearerAuth: []
      tags:
        - Booth
      summary: Get an existing booth
      description: Get an existing booth by Id
      operationId: getBooth
      parameters:
        - name: id
          description: ID to Get booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'booth'
          description: |
            List of populatable columns:
              - booth_images.booth_template_position
              - booth_images.image
              - booth_projects.booth_template_position
              - booth_projects.project
              - booth_products.booth_template_position
              - booth_products.product
              - booth_videos.booth_template_position
              - booth_videos.video
              - booth_template
              - location
              - location.space_template_location
              - exhibition
              - meetings
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booth'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    delete:
      security:
      - bearerAuth: []
      tags:
        - Booth
      summary: Delete an existing booth
      description: Delete an existing booth by Id
      operationId: deleteBooth
      parameters:
        - name: id
          description: ID to delete booth
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
...
  '/users':
    get:
      tags:
        - User
      summary: Show list users
      description: Show list users
      operationId: getUsers
      parameters:
      - in: query
        name: page
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '1'
        description: The number of items to skip before starting to collect the result set
      - in: query
        name: limit
        schema:
          type: string
          pattern: '^[0-9]+'
          example: '10'
        description: The numbers of items to return
      - in: query
        name: search
        schema:
          type: string
          example: 'image'
        description:
         " Searching on allow columns:['id', 'name'', 'created_at']\n
          example: http://localhost:3000/api/v1/users?search=ABC"
      - in: query
        name: sort_by
        allowReserved: true
        schema:
          type: string
          example: 'mime:DESC'
        description: "Sort on allow columns:[['id', 'name', 'created_by','created_date','role','created_at'],[DESC,ASC]]\n default column:['created_at', 'DESC']"
      - in: query
        name: filter
        allowReserved: true
        schema:
          type: array
          items:
            type: string
            example: 
              - status:$eq:active'
              - created_date:$eq:active'
        description: Filter by the field with syntax `column_name:operator:value`.
      - in: query
        name: withDeleted
        schema:
          type: string
          example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    post:
      security:
      - bearerAuth: []
      tags:
        - User
      summary: Add a new user
      description: Add a new user
      operationId: addUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/User'
              required:
                - password
                - name
                - phone
                - email
                - role_id
                - status
                - enterprise_id
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/User'
                required:
                  - password
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/users/{id}':
    patch:
      tags:
        - User
      security:
      - bearerAuth: []
      summary: Update user
      description: Update user
      operationId: UserUpdate
      parameters:
      - name: id
        description: ID to update user
        in: path
        required: true
        schema:
          type: string
          pattern: '^[0-9]+$'
      requestBody:
        description: update user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    get:
      security:
      - bearerAuth: []
      tags:
        - User
      summary: Get user  by ID
      description: Return a single user
      operationId: getUsersById
      parameters:
        - name: id
          description: ID to get user
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
        - in: query
          name: populate
          schema:
            type: array
            items:
              type: string
              example: 'role'
          description: |
            - role
        - in: query
          name: withDeleted
          schema:
            type: string
            example: 'true'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/User'
                required:
                  - id
                  - name
                  - phone
                  - email
                  - password
                  - role_id
                  - status
                  - enterprise_id
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
    delete:
      security:
      - bearerAuth: []
      tags:
        - User
      summary: Delete an existing user
      description: Delete an existing user by Id
      operationId: deleteUser
      parameters:
        - name: id
          description: ID to delete user
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]+$'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Not Found
                    details:
                      - field: .params.id
                        message: 'The id not exist: "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
  '/users/change-password':
    post:
      tags:
        - User
      security:
      - bearerAuth: []
      summary: Change password for user
      description: Change password for user
      operationId: userChangePassword
      requestBody:
        description: update user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
        required: true
      responses:
        '204':
          description: No Content
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Bad Request
                    details:
                      - field: .params.id
                        message: 'should match pattern "^[0-9]+$"'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseBody'
              examples:
                example-1:
                  value:
                    message: Internal server error
                    details: []
...
components:
  schemas:
...
    ErrorResponseBody:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
    Booth:
      type: object
      required:
        - id
        - name
        - enterprise_id
        - created_by
        - location_id
        - booth_template_id
      properties:
        id:
          type: integer
          example: 1
          readOnly: true
        name:
          type: string
        enterprise_id:
          type: integer
          example: 1
        created_by:
          type: integer
          readOnly: true
          example: 1
        created_date:
          type: string
          readOnly: true
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-01-01T00:00:00.000Z'
        location_id:
          type: integer
          writeOnly: true
          example: 1
        booth_template_id:
          type: integer
          writeOnly: true
          example: 1
        live_streams:
          type: array
          items:
            $ref: '#/components/schemas/LiveStream'
        booth_images:
          type: array
          items:
            $ref: '#/components/schemas/BoothImage'
        booth_videos:
          type: array
          items:
            $ref: '#/components/schemas/BoothVideo'
        booth_projects:
          type: array
          items:
            $ref: '#/components/schemas/BoothProject'
        booth_products:
          type: array
          items:
            $ref: '#/components/schemas/BoothProduct'
...
    User:
        type: object
        required:
          - id
        properties:
          id:
            type: integer
            readOnly: true
          name:
            type: string
          phone:
            type: string
            pattern: '^(((\+|)84)|0)([0-9]{9})$'
            example: 0707060142
          email:
              type: string
              format: email
          password:
            type: string
            format: password
            writeOnly: true
            minLength: 6
            maxLength: 255
          role_id:
            type: number
            writeOnly: true
          role:
            $ref: '#/components/schemas/Role'
          status:
            type: string
            enum: [active, inactive]
          created_by:
            type: integer
            readOnly: true
          created_date:
            type: string
            readOnly: true
          enterprise_id:
            type: number
    Role:
      readOnly: true
      required:
        - id
      properties:
        id:
          type: integer
        name:
          type: string
    Password:
        type: object
        required:
          - current_password
          - password
        properties:
          current_password:
            type: string
            format: password
            writeOnly: true
            minLength: 6
            maxLength: 255
          password:
            type: string
            format: password
            writeOnly: true
            minLength: 6
            maxLength: 255
...
    LiveStream:
      type: object
      required:
        - id
        - link
        - time
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        link:
          type: string
          minLength: 1
          example: 'www.goole.com'
        time:
          type: string
          minLength: 1
          pattern: '^[1-9]\d{3}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}.\d{3}Z$'
          example: '2022-12-01T00:00:00.000Z'
    BoothImage:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothVideo:
      type: object
      required:
        - id
        - booth_template_position_id
        - video_id
      properties:
        id:
          type: integer
          readOnly: true
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video selected from data center
          example: 1
        video_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          example: 1
    BoothProject:
      type: object
      required:
        - id
        - booth_template_position_id
        - image_id
        - title
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        view:
          type: integer
          readOnly: true
          example: 1
        like:
          type: integer
          readOnly: true
          example: 1
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if video or image selected from data center
        image_id:
          type: integer
          readOnly: true
        media_data:
          type: string
          writeOnly: true
          description: This column exists if video or image selected from your local
          example: 'base64 string'
        title:
          type: string
          maxLength: 255
          example: 'Tên dự án'
        description:
          type: string
          maxLength: 20000
          example: 'Nội dung mô tả cho dự án'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
    BoothProduct:
      type: object
      required:
        - id
        - name
        - price
        - purchase_link
        - booth_template_position_id
        - description
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          example: 'mận Hà Nội'
        price:
          type: number
          format: float
          example: 35000
        purchase_link:
          type: string
          maxLength: 255
          example: 'https://link_mua_san_pham/ma_ha_noi.png'
        description:
          type: string
          maxLength: 500
          example: 'Mô tả sản phẩm'
        selected_media_id:
          type: integer
          writeOnly: true
          description: This column exists if image selected from data center
          example: 1
        image_id:
          type: integer
          readOnly: true
          example: 1
        media_data:
          type: string
          writeOnly: true
          description: This column exists if image selected from your local
          example: 'base64 string'
        booth_template_position_id:
          type: integer
          writeOnly: true
          example: 1
...
    PaginatedUsers:
      type: object
      required:
        - page
        - limit
        - total
        - users
      properties:
        page:
          type: integer
          readOnly: true
          example: 1
        limit:
          type: integer
          readOnly: true
          example: 1
        total:
          type: integer
          readOnly: true
          example: 1
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT